updated calculator.h

+++ calculator.h
@@ -9,9 +9,9 @@ double add(double a, double b);
 double subtract(double a, double b);
 double multiply(double a, double b);
 double divide(double a, double b);
-double lcm(double a, double b);
+int lcm(double a, double b);
 double factorial(double n);
-double gcd(double a, double b);
+int gcd(double a, double b);
 int randomnumber(int a, int b); 
 
 #endif
